Remove dependency on possibly stale annotation objects in AnnotationUISync
When an annotation is selected in the page, AnnotationUISync would
receive a message containing the annotation's local tag and then look up
the annotation in a (tag <=> Annotation) map maintained by AnnotationSync
in order to get its ID, that could then be passed to
`annotationUI.selectAnnotations()` in order to set the selection.

When an annotation edit is saved, the instance of the annotation is replaced
by an updated version in the app's state store and consequently
AnnotationSync has an outdated version in its cache. In the case of a
new annotation, the stale copy in the cache does not have an ID.

This commit avoids the problem by allowing the annotationUI functions
for selecting and focusing annotations to simply take a list of IDs _or_
tags rather than requiring actual Annotation instances.