Update client's document data extraction
The `annotation.uri` now always is a linkable URI if we got that data
when the annotation got created. There is no need anymore to try and
find another linkable URI from the document data.
This also removes the type check of `annotation.document.title` since
the presentation layer on the backend always returns an array of
strings (or the `title` property is missing).